Add SCSI in NCCFV for TD guest
The FV in `IntelTdx/IntelTdxX64.fdf` was split into two FVs, one for a
TD guest and one for non-cc guest. The SCSI driver was moved to the
non-cc NCCFV. In order for PVE to use an image with SCSI include the
SCSI driver in the FV for TD guest.

Signed-off-by: Philipp Giersfeld <philipp.giersfeld@canarybit.eu>
Signed-off-by: Anton Iacobaeus <anton.iacobaeus@canarybit.eu>
Link: https://lore.proxmox.com/20251028125459.287308-4-anton.iacobaeus@canarybit.eu
[FE: rebase on latest master]
Signed-off-by: Fiona Ebner <f.ebner@proxmox.com>